The rent window has been added and its logic has been written

+++ src/main/java/com/asdeire/domain/service/impl/SignUpService.java
@@ -41,7 +41,7 @@ final class SignUpService {
     public void signUp(String username,
                        String password,
                        String email,
-                       int balance,
+                       double balance,
                        Supplier<String> waitForUserInput) {
         try {
             String verificationCode = generateAndSendVerificationCode(email);
